Fix 4k transfer sizes.
If the LSB of wLength happens to be zero, then the logic to determine
the transfer size goes wrong, resulting in a missing DATA phase. This
seems to be caused by the final byte of the wLength being latched in
on the same clock that also does the handling of the SETUP packet.

To resolve this, we add another state (SETUP_DONE) that just delays
the processing of the SETUP packet until all bytes of wLength have
settled.