Revert to DFU v1.1, it's a little simpler.
The official DFU standard relies purely on the use of a zero-length
packet to identify the end of the firmware file, while DfuSe (1.1a)
expects there to be more commands implemented for addressing.

We also fix a bug in which the DFU endpoint would fail to send a
zero-length-packet at the end of firmware readout.