Fixup SB_RAM40_4K instance.
Setting the READ_MODE and WRITE_MODE explicitly to 512x8 doesn't
seem to actually work. So leave the parameters at their defaults and
just use 256x16 mode, which seems to work fine.

We are still experiencing a bug where there's an extra byte being
inserted at the start of the firmware write, but I suspect that its
more to do with the USB flow control behavior being kinda nuts.