Fixup DATA_OUT -> STATUS_IN transition bug.
Since out_ep_acked only pulses high at the start of a packet, and
not once the buffer has been handled, it doesn't give us a good
signal on when to terminate the DATA_OUT state. This cleans up
some hacks that were processing data packets outside of the DATA_OUT
state.