Fix address latching in spiflash bridge.
The read and write addresses were being latched incorrectly, leading to
erratic behaviour. I think the cause was that the command setup is done
in async logic while the address latch was done synchronously, which
led to a timing race between the address latch and the command setup.

To resolve the situation, we just remove all address latching from the
spiflash bridge, and instead expect the user of the module to keep the
address lines stable for the duration of a read or write request.

It now looks like we are able to write the flash correctly over DFU,
but we are now getting corrupted results when attempting a read. The
fun never stops...