Add WRITE_EOF state to the flash state machine.
Without an EOF state, we never release the chip select at the end of
a write, leading to a couple of extra junk bytes wrapping around and
getting written to the start of each page. Adding an EOF state allows
the flash to correctly deactivate the chipselect.

Also... IT WORKS!