Add latch delay for write data valid signal.
There seems to be a one cycle delay from the rising edge of the
out_ep_data_avail signal to when the data is actually presented on
the out_ep_data bus. To get the correct data, we need to delay the
processing of the data by a cycle as well.