Check boot switch on POR and reboot if not pressed.
If the device powers on with the boot switch pressed, the device should
remain in DFU mode, othewise it should boot into the application image.
On the ECP5, this is done by asserting the PROGRAMN pin low, but we
forgot to connect PROGRAMN to any FPGA-controlled pin, so it has been
bodged onto ETH_RESET (ball C17) for the prototype boards.

This doesn't actually multiboot anything for now, it will just result
in a boot loop until the boot pin is asserted.