Move USB PHY into its own module and add DFU reset handling.
The DFU spec uses a USB reset condition to transition back and forth
between DFU and Application mode, but the integration of the PHY into
the USB stack makes it difficult to detect and signal this event up
to the top where multibooting is handled.

We also wound up with some I2C work-in-progress landing in this commit
because it was a pain to separate out.