Initial support for Arty-A7
This work is based off of a pull request to add the Digilent Arty-A7
board from https://github.com/riktw/tinydfu-bootloader:master and
has been modified to build with as a TCL script to keep Vivado cruft
to a minimum, and also targets the 100T-size parts (because that's
all I happened to have).

This seems to enumerate by USB just fine, but there is a problem that
any writes passed 512kB fail. This seems to be due to a weird flash
chip that Digilent used, and only supports the 4kB erase command at
the start of the chip.